refactor script back as a service, add full client
